- Added test for parser; - Added response on parser controller; - Test on controller no nedd to write, user EDostavkaParserTest.java for testing;

+++ src/main/java/com/bareisha/producthunter/core/dto/ProductDto.java
@@ -2,14 +2,12 @@ package com.bareisha.producthunter.core.dto;
 
 import lombok.Data;
 
-import java.math.BigDecimal;
-
 @Data
 public class ProductDto {
 
     private String title;
-    private BigDecimal price;
-    private BigDecimal priceDiscount;
+    private Double price;
+    private Double priceDiscount;
     private String country;
 
 }
